move pipeline loading to a separate file
